Feature/stats system (#31)
* Stats Scriptable Object y add player stats

* Refactor enemies scripts

* Scene save

* Stats on enemies

* Stats UI

+++ TPF/Assets/Scripts/UI/HealthUI.cs
@@ -34,7 +34,7 @@ public class HealthUI : MonoBehaviour
 
     }
 
-    void UpdateHealthDisplay(int newHealth)
+    void UpdateHealthDisplay(float newHealth)
     {
         slider.value = newHealth;
     }
